Update render.yml
Output showed weird bugs about fontist not being installed, which is an issue of one of the dependency `fontist` in the action.

It [turns out](https://rubygems.org/gems/fontist/versions) that fontist v2.0.0 was recently published but yanked, and setup-fontist tries to install the latest version despite having been yanked.

We fix this by pinning to 1.x.