Remove wrong upper bound on Best-Fit
I erroneously stated and assumed an upper bound on the absolute random-order-ratio of Best-Fit of 1.5, which is actually not known in the literature, let alone in the one I erroneously cited.

The best-known upper bound is only 1.7, which is a simple corollary from the upper-bound of 1.7 of the absolute approximation-ratio of Best-Fit.